// packages/civ7-map-policy/src/map-metadata.gen.ts
export const CIV7_MAP_INFO_COLUMN_DESCRIPTORS = Object.freeze([
  Object.freeze({
    name: "MapSizeType",
    sqlType: "TEXT",
    nullable: false,
    hasDefault: false,
    defaultValue: null,
  } as const),
  Object.freeze({
    name: "AllOnLargestLandmass",
    sqlType: "BOOLEAN",
    nullable: false,
    hasDefault: true,
    defaultValue: false,
  } as const),
  Object.freeze({
    name: "Continents",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 1,
  } as const),
  Object.freeze({
    name: "DefaultPlayers",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "Description",
    sqlType: "TEXT",
    nullable: true,
    hasDefault: false,
    defaultValue: null,
  } as const),
  Object.freeze({
    name: "GridHeight",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "GridWidth",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "LakeGenerationFrequency",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 25,
  } as const),
  Object.freeze({
    name: "LakeSizeCutoff",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "Name",
    sqlType: "TEXT",
    nullable: false,
    hasDefault: false,
    defaultValue: null,
  } as const),
  Object.freeze({
    name: "NumNaturalWonders",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "OceanWidth",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "PlayersLandmass1",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "PlayersLandmass2",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 0,
  } as const),
  Object.freeze({
    name: "StartSectorCols",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 3,
  } as const),
  Object.freeze({
    name: "StartSectorRows",
    sqlType: "INTEGER",
    nullable: false,
    hasDefault: true,
    defaultValue: 3,
  } as const),
] as const);
// ...
export const CIV7_STANDARD_MAP_INFO_ROWS = Object.freeze([
  Object.freeze({
    MapSizeType: "MAPSIZE_TINY",
    AllOnLargestLandmass: false,
    Continents: 2,
    DefaultPlayers: 4,
    Description: "LOC_MAPSIZE_TINY_DESCRIPTION",
    GridHeight: 38,
    GridWidth: 60,
    LakeGenerationFrequency: 25,
    LakeSizeCutoff: 6,
    Name: "LOC_MAPSIZE_TINY_NAME",
    NumNaturalWonders: 3,
    OceanWidth: 4,
    PlayersLandmass1: 3,
    PlayersLandmass2: 1,
    StartSectorCols: 2,
    StartSectorRows: 3,
  } as const),
  Object.freeze({
    MapSizeType: "MAPSIZE_SMALL",
    AllOnLargestLandmass: false,
    Continents: 4,
    DefaultPlayers: 6,
    Description: "LOC_MAPSIZE_SMALL_DESCRIPTION",
    GridHeight: 46,
    GridWidth: 74,
    LakeGenerationFrequency: 25,
    LakeSizeCutoff: 8,
    Name: "LOC_MAPSIZE_SMALL_NAME",
    NumNaturalWonders: 4,
    OceanWidth: 4,
    PlayersLandmass1: 4,
    PlayersLandmass2: 2,
    StartSectorCols: 3,
    StartSectorRows: 3,
  } as const),
  Object.freeze({
    MapSizeType: "MAPSIZE_STANDARD",
    AllOnLargestLandmass: false,
    Continents: 4,
    DefaultPlayers: 8,
    Description: "LOC_MAPSIZE_STANDARD_DESCRIPTION",
    GridHeight: 54,
    GridWidth: 84,
    LakeGenerationFrequency: 25,
    LakeSizeCutoff: 8,
    Name: "LOC_MAPSIZE_STANDARD_NAME",
    NumNaturalWonders: 5,
    OceanWidth: 4,
    PlayersLandmass1: 5,
    PlayersLandmass2: 3,
    StartSectorCols: 3,
    StartSectorRows: 4,
  } as const),
  Object.freeze({
    MapSizeType: "MAPSIZE_LARGE",
    AllOnLargestLandmass: false,
    Continents: 6,
    DefaultPlayers: 10,
    Description: "LOC_MAPSIZE_LARGE_DESCRIPTION",
    GridHeight: 60,
    GridWidth: 96,
    LakeGenerationFrequency: 25,
    LakeSizeCutoff: 10,
    Name: "LOC_MAPSIZE_LARGE_NAME",
    NumNaturalWonders: 6,
    OceanWidth: 8,
    PlayersLandmass1: 6,
    PlayersLandmass2: 4,
    StartSectorCols: 3,
    StartSectorRows: 4,
  } as const),
  Object.freeze({
    MapSizeType: "MAPSIZE_HUGE",
    AllOnLargestLandmass: false,
    Continents: 6,
    DefaultPlayers: 10,
    Description: "LOC_MAPSIZE_HUGE_DESCRIPTION",
    GridHeight: 66,
    GridWidth: 106,
    LakeGenerationFrequency: 25,
    LakeSizeCutoff: 10,
    Name: "LOC_MAPSIZE_HUGE_NAME",
    NumNaturalWonders: 7,
    OceanWidth: 8,
    PlayersLandmass1: 6,
    PlayersLandmass2: 6,
    StartSectorCols: 3,
    StartSectorRows: 4,
  } as const),
] as const);